refactor: extract conversation message list
